game-engine/requirements.txt
# Core dependencies
google-genai>=1.0.0
python-dotenv>=1.0.0
pygame>=2.5.0

# Web framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
jinja2>=3.1.3
python-multipart>=0.0.9

# Data handling
pydantic>=2.0.0
